Established thread for lighting

+++ PixelGameEngineTinker/PixelGameEngineTinker/Light.h
@@ -5,84 +5,84 @@
 class Light
 {
 private:
-	std::int16_t _r;
-	std::int16_t _g;
-	std::int16_t _b;
-	std::int16_t _a;
+	std::int32_t _r;
+	std::int32_t _g;
+	std::int32_t _b;
+	std::int32_t _a;
 
 public:
 	// Constructors
 	Light() : _r( 0 ), _g( 0 ), _b( 0 ), _a( 0 ) {}
-	Light( std::int16_t r_, std::int16_t g_, std::int16_t b_ ) : _r( r_ ), _g( g_ ), _b( b_ ), _a( 0 ) {}
-	Light( std::int16_t r_, std::int16_t g_, std::int16_t b_, std::int16_t a_ ) : _r( r_ ), _g( g_ ), _b( b_ ), _a( a_ ) {}
+	Light( std::int32_t r_, std::int32_t g_, std::int32_t b_ ) : _r( r_ ), _g( g_ ), _b( b_ ), _a( 0 ) {}
+	Light( std::int32_t r_, std::int32_t g_, std::int32_t b_, std::int32_t a_ ) : _r( r_ ), _g( g_ ), _b( b_ ), _a( a_ ) {}
 
 
 	// Getters
-	std::int16_t getRed() const
+	std::int32_t getRed() const
 	{
 		return this->_r;
 	}
 
-	std::int16_t getGreen() const
+	std::int32_t getGreen() const
 	{
 		return this->_g;
 	}
 
-	std::int16_t getBlue() const
+	std::int32_t getBlue() const
 	{
 		return this->_b;
 	}
 
-	std::int16_t getAlpha() const
+	std::int32_t getAlpha() const
 	{
 		return this->_a;
 	}
 
 
 	// Setters
-	void setRed( std::int16_t r_ )
+	void setRed( std::int32_t r_ )
 	{
 		this->_r = r_;
 		return;
 	}
 
-	void setGreen( std::int16_t g_ )
+	void setGreen( std::int32_t g_ )
 	{
 		this->_g = g_;
 		return;
 	}
 
-	void setBlue( std::int16_t b_ )
+	void setBlue( std::int32_t b_ )
 	{
 		this->_b = b_;
 		return;
 	}
 
-	void setAlpha( std::int16_t a_ )
+	void setAlpha( std::int32_t a_ )
 	{
 		this->_a = a_;
 		return;
 	}
 
-	void addRed( std::int16_t r_ )
+	void addRed( std::int32_t r_ )
 	{
 		this->_r += r_;
 		return;
 	}
 
-	void addGreen( std::int16_t g_ )
+	void addGreen( std::int32_t g_ )
 	{
 		this->_g += g_;
 		return;
 	}
 
-	void addBlue( std::int16_t b_ )
+	void addBlue( std::int32_t b_ )
 	{
 		this->_b += b_;
 		return;
 	}
 
-	void addAlpha( std::int16_t a_ )
+	void addAlpha( std::int32_t a_ )
 	{
 		this->_a += a_;
 		return;
